modified:   workspaces/DummyRosterRC1/DummyRoster.Common.EntityModel/Models/Invoice.cs

+++ workspaces/DummyRosterRC1/DummyRoster.Common.EntityModel/Models/Invoice.cs
@@ -15,10 +15,10 @@ public partial class Invoice
   public int Id { get; set; }
 
   [Column(TypeName = "INT")]
-  public long? FormId { get; set; }
+  public int? FormId { get; set; }
 
   [Column(TypeName = "INT")]
-  public long? ProductId { get; set; }
+  public int? ProductId { get; set; }
 
   public string? Note { get; set; }
 
@@ -26,7 +26,7 @@ public partial class Invoice
   public byte[] UnitPrice { get; set; } = null!;
 
   [Column(TypeName = "SMALLINT")]
-  public long Quantity { get; set; }
+  public int Quantity { get; set; }
 
   [Column(TypeName = "NUMERIC")]
   public byte[] PriceCut { get; set; } = null!;
